Dramatic rewrite of map matching system.
This system matches maps and their "fingerprints" to devices and their available events. If a map matches, a symlink is made in /etc/rejoy/maps and xboxdrv is started. The list of map candidates is determined by the driver in use. Sony's DS3 is very exotic, and it is matched first before the more standard DS4. So far, the only hid-generic map I have is for wired 3rd party PS3 and PS4 controllers. I have a Mayflash PS2 adapter that I will add. I would also like to make mappings for Wii controllers. That should cover the majority use cases. For the rest, I intend to create a fallback that converts a SDL mapping to an Xboxdrv mapping. This is possible as long as the controller we're converting for is actually there to grab event codes from; the order of the event codes is the same order as button and axis indexes. Axes 16-23 are always hats.